examples/ui/gradio/gradio-tgi-multi-model-rag-redis/requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu118/
aiofiles==23.2.1; python_version >= '3.7'
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
altair==5.1.2; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cmake==3.27.6
contourpy==1.1.1; python_version >= '3.8'
cycler==0.12.1; python_version >= '3.8'
dataclasses-json==0.6.1; python_version >= '3.7' and python_version < '4.0'
fastapi==0.103.2; python_version >= '3.7'
ffmpy==0.3.1
filelock==3.12.4; python_version >= '3.8'
fonttools==4.43.1; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
fsspec==2023.9.2; python_version >= '3.8'
greenlet==3.0.0; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==0.18.0; python_version >= '3.8'
httpx==0.25.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
importlib-resources==6.1.0; python_version >= '3.8'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonschema==4.19.1; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
kiwisolver==1.4.5; python_version >= '3.7'
lit==17.0.6
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
matplotlib==3.8.0; python_version >= '3.9'
mpmath==1.3.0
markdown==3.5; python_version >= '3.8'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
networkx==3.1; python_version >= '3.8'
nltk==3.8.1; python_version >= '3.7'
numpy==1.26.0; python_version < '3.13' and python_version >= '3.9'

packaging==23.2; python_version >= '3.7'
pandas==2.1.1; python_version >= '3.9'
pillow==10.0.1; python_version >= '3.8'
pydantic==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
pydub==0.25.1
pdfkit==1.0.0; python_version >= '3.8'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.1; python_version >= '3.7'
referencing==0.30.2; python_version >= '3.8'
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.10.6; python_version >= '3.8'
scikit-learn==1.3.1; python_version >= '3.8'
scipy==1.11.3; python_version < '3.13' and python_version >= '3.9'
semantic-version==2.10.0; python_version >= '2.7'
sentencepiece==0.1.99
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.21; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
sympy==1.12; python_version >= '3.8'
tenacity==8.2.3; python_version >= '3.7'
threadpoolctl==3.2.0; python_version >= '3.8'
toolz==0.12.0; python_version >= '3.5'
torch==2.0.1; python_full_version >= '3.8.0'
torchvision==0.15.2; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
triton==2.0.0; platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.8.0; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2023.3; python_version >= '2'
urllib3==2.0.6; python_version >= '3.7'
uvicorn==0.23.2; python_version >= '3.8'
websockets==11.0.3; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
prometheus-client==0.17.1; python_version >= '3.8'
itsdangerous==2.1.2
text_generation==0.6.1; python_version >= '3.8'
faiss-cpu
pgvector==0.2.5; python_version >= '3.8'
psycopg[binary,pool]==3.1.18; python_version >= '3.8'
langchain==0.2.16; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.2.16; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.2.38; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-elasticsearch==0.2.2; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.2.4; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.75; python_version < '4.0' and python_full_version >= '3.8.1'
orjson==3.9.14; python_version >= '3.8'
langchain-openai==0.1.19; python_version >= '3.8'
gradio==4.44.1
gradio-client==1.3.0
huggingface-hub==0.28.1 
python-multipart==0.0.20 
ruff==0.9.6
safetensors==0.5.2
sentence-transformers==3.4.1
tokenizers==0.21.0
transformers==4.48.3
